pyproject.toml
[project]
name = "westsmith-co-uk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "flask>=3.1.3",
    "markdown>=3.10.2",
    "pre-commit>=4.2.0",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
]
